Remove babel and test
Babel was included as a dependency in order to support the unit tests in
this repo. Unfortunately, it looks like there aren't any real unit tests
in this repo, making babel unnecessary. Babel has been known to cause
issues in ES6 modules, making it a somewhat unhelpful dependency for my
use case.